+++ src/plots/3_body_chaos.png
