+++ src/plots/BUTTERFLY IV.png
