+++ src/plots/orbits_2light_1massive.png
